# backend/build/contracts.openapi.json
{
  "components": {
    "schemas": {
...
      "AgentObservability": {
        "additionalProperties": false,
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentRecord"
            },
            "title": "Agents",
            "type": "array"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/SystemComponent"
            },
            "title": "Components",
            "type": "array"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolRecord"
            },
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "agents",
          "tools",
          "components"
        ],
        "title": "AgentObservability",
        "type": "object"
      },
...
      "DataMode": {
        "enum": [
          "illustrative_fixture"
        ],
        "title": "DataMode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PresentationEnvelope_AgentObservability_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AgentObservability"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[AgentObservability]",
        "type": "object"
      },
      "PresentationEnvelope_AgentRecord_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AgentRecord"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[AgentRecord]",
        "type": "object"
      },
      "PresentationEnvelope_AlternativeCollection_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AlternativeCollection"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[AlternativeCollection]",
        "type": "object"
      },
      "PresentationEnvelope_AlternativeSession_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AlternativeSession"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[AlternativeSession]",
        "type": "object"
      },
      "PresentationEnvelope_DecisionCollection_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DecisionCollection"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[DecisionCollection]",
        "type": "object"
      },
      "PresentationEnvelope_Governance_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Governance"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[Governance]",
        "type": "object"
      },
      "PresentationEnvelope_NewsCollection_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NewsCollection"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[NewsCollection]",
        "type": "object"
      },
      "PresentationEnvelope_Overview_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Overview"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[Overview]",
        "type": "object"
      },
      "PresentationEnvelope_Portfolio_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Portfolio"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[Portfolio]",
        "type": "object"
      },
      "PresentationEnvelope_StoryDetail_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StoryDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[StoryDetail]",
        "type": "object"
      },
      "PresentationEnvelope_WeeklySummary_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WeeklySummary"
          },
          "meta": {
            "$ref": "#/components/schemas/PresentationMeta"
          }
        },
        "required": [
          "meta",
          "data"
        ],
        "title": "PresentationEnvelope[WeeklySummary]",
        "type": "object"
      },
      "PresentationMeta": {
        "additionalProperties": false,
        "properties": {
          "asOf": {
            "format": "date-time",
            "title": "Asof",
            "type": "string"
          },
          "dataMode": {
            "$ref": "#/components/schemas/DataMode",
            "default": "illustrative_fixture"
          },
          "fixtureVersion": {
            "const": "prism-demo-v1",
            "default": "prism-demo-v1",
            "title": "Fixtureversion",
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "title": "Generatedat",
            "type": "string"
          },
          "range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DateRange"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "generatedAt",
          "asOf"
        ],
        "title": "PresentationMeta",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "PRISM: One signal. Multiple perspectives. Better decisions. Paper-only infrastructure API; execution is disabled by default.",
    "title": "PRISM API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/presentation/agents": {
      "get": {
        "operationId": "agents_api_v1_presentation_agents_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_AgentObservability_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agents",
        "tags": [
          "presentation"
        ]
      }
    },
    "/api/v1/presentation/agents/{agent_id}": {
      "get": {
        "operationId": "agent_api_v1_presentation_agents__agent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_AgentRecord_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent",
        "tags": [
          "presentation"
        ]
      }
    },
    "/api/v1/presentation/alternatives": {
      "get": {
        "operationId": "alternatives_api_v1_presentation_alternatives_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_AlternativeCollection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alternatives",
        "tags": [
          "presentation"
        ]
      }
    },
    "/api/v1/presentation/alternatives/{session_id}": {
      "get": {
        "operationId": "alternative_api_v1_presentation_alternatives__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_AlternativeSession_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alternative",
        "tags": [
          "presentation"
        ]
      }
    },
    "/api/v1/presentation/decisions": {
      "get": {
        "operationId": "decisions_api_v1_presentation_decisions_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_DecisionCollection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decisions",
        "tags": [
          "presentation"
        ]
      }
    },
    "/api/v1/presentation/decisions/{decision_id}": {
      "get": {
        "operationId": "decision_api_v1_presentation_decisions__decision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "title": "Decision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_StoryDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decision",
        "tags": [
          "presentation"
        ]
      }
    },
    "/api/v1/presentation/governance": {
      "get": {
        "operationId": "governance_api_v1_presentation_governance_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_Governance_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Governance",
        "tags": [
          "presentation"
        ]
      }
    },
    "/api/v1/presentation/news": {
      "get": {
        "operationId": "news_api_v1_presentation_news_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "in": "query",
            "name": "significance",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Significance"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_NewsCollection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "News",
        "tags": [
          "presentation"
        ]
      }
    },
    "/api/v1/presentation/overview": {
      "get": {
        "operationId": "overview_api_v1_presentation_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_Overview_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview",
        "tags": [
          "presentation"
        ]
      }
    },
    "/api/v1/presentation/portfolio": {
      "get": {
        "operationId": "portfolio_api_v1_presentation_portfolio_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_Portfolio_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Portfolio",
        "tags": [
          "presentation"
        ]
      }
    },
    "/api/v1/presentation/weekly-summary": {
      "get": {
        "operationId": "weekly_summary_api_v1_presentation_weekly_summary_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "prism_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prism Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationEnvelope_WeeklySummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Weekly Summary",
        "tags": [
          "presentation"
        ]
      }
    },
...
  }
}